[compiler] Reduce wb verification performance overhead
This CL adds a fast-path check to skipped write barrier verification
written in inline-assembly to avoid the cost of saving/restoring
registers and the C++ call.

The new method PreCheckSkippedWriteBarrier jumps directly back to the
regular code path once it determines that skipping the write barrier
was safe. This method doesn't need to be exhaustive and is optional
because we still use the C++ method as a fallback and for error
reporting.

This CL reduces the overhead from -12% to -3.3% on Speedometer3 and
-26.7% to -4% on Jetstream2 (both on M1).

Bug: 437096305
Change-Id: I7c280f6025a8204b85cb7f7a474162902b074df8
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6929591
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102512}