[wasm-custom-desc] JS Interop: wrap methods
instead of "stamping" them.
The problem we need to solve is that methods installed on prototypes
should not ignore the JS-side "receiver" (as regular exported Wasm
functions do), but rather pass it on as the first Wasm-side parameter.
The previous approach was to set a bit on the exported function to
indicate the changed behavior; this CL uses wrapper functions instead,
at the cost of slower instantiation.

Bug: 403372470
Change-Id: I132e4e5f053de7dc5737b7edaa3b97507084b2c6
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6948076
Commit-Queue: Jakob Kummerow <jkummerow@chromium.org>
Reviewed-by: Daniel Lehmann <dlehmann@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102516}