[heap] Remove bogus CHECK for contiguous RO space
The check wanted to make sure that the red zone start is always
aligned to the pointer compression base alignment.

This is not correct in the case we have a red zone that's at the very
beginning of the reservation (and has an overlapping range of
kContiguousReadOnlyReservationSize).

Bug: 429538831
Change-Id: Idfcd736e513806f92ff5ca6ba9f62b0e77cec24b
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6953761
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Dominik Inführ <dinfuehr@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Auto-Submit: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102518}