[turbolev] Do not try to guess map from InlinedAllocation later on
Because this won't be aware of potential transitions that happened
since then.
Also, this shouldn't be necessary in the vast majority of cases,
since when creating an InlinedAllocation, we almost always call
BuildStoreMap afterwards, which will take care of recording the map
in the KNA.

This fixes a deopt loop in Jetstream2/raytrace with --turbolev-future,
which improves Average-Score by 75% and Total-Score by 35% (on my x64
workstation).

Bug: 431933185
Change-Id: I5dce98093f958d195de7aa424aa149c2b9aacfd8
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6959389
Reviewed-by: Victor Gomes <victorgomes@chromium.org>
Auto-Submit: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102555}