[maglev] Fix register liveness for lazy deopts with exception handlers
Lazy deopt frames for operations that can throw did not consider the
liveness of registers at the corresponding exception handler. This
could lead to registers being marked as dead, even though they would
be needed if an exception was caught.

This patch fixes this by intersecting the current liveness with the
exception handler's liveness when building the deopt frame for a
potentially throwing operation.

This also means removing the special handling which skips result
registers when iterating lazy deopt inputs. This is ok in the common
case, since we already mark them as dead (and therefore not in the
compact IFS) if not needed for a catch.

Change-Id: I1fd0f3bd1ef429be2965941d953d60ef90e695ad
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6959220
Auto-Submit: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Victor Gomes <victorgomes@chromium.org>
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102580}