[riscv] Only compile vlen intrinsic for RVV targets
The '__riscv_vlenb' intrinsic must not be used if the compiler doesn't
compile for a CPU with 'RVV' support. Instead of falling back to
assembler, we mark the 'vlen_intrinsic' function with the 'target'
attribute, and make sure to only call it if RVV is present.

Change-Id: Ie10b105161df1134c93dde87a65571537c1215d4
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6959231
Commit-Queue: Florian Loitsch <floitsch@rivosinc.com>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Kasper Lund <kasperl@rivosinc.com>
Reviewed-by: Ji Qiu <qiuji@iscas.ac.cn>
Cr-Commit-Position: refs/heads/main@{#102593}