Release store of Code to JSDispatchTable
This fixes a race where a Maglev thread was creating a Code object and
storing it in the JSDispatchTable, then a Turbofan thread was reading
it, and it was reading its inlined_bytecode_size field, but it was
uninitialized because it wasn't seeing the store of this field yet.

Note that without leap tiering, store of the Code object into the
JSFunction object was already a release store.

Bug: 441515240
Change-Id: I9df662256c2e4e889ecd496b5817271b7586bfad
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6954065
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102619}