[js-objects] Handle TheHole in JSCreateObject fast path
The `prototype_or_initial_map` slot on a constructor can be
`TheHole`. The fast path in `JSCreateObject` did not handle this
case and would fail on the subsequent `Cast<Map>`. This change adds
a check and redirects to the slow path when the slot is `TheHole`.

Fixed: 444622225
Change-Id: Ic1d5b4a64fb52017bba920f4959f56613bddec72
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6965651
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102631}