[heap] Relocate holes to an unmappable page
This patch relocates all hole values to a dedicated memory page within
the WasmNull object's padding. This allows for unmapping the page
containing the holes as a security hardening measure, controlled by the
new experimental flag `--unmap-holes`.

The WasmNull is extended to guarantee to span two OS pages, so that the
first page after the header can continue to be unmapped for Wasm's
purposes, and the second page can be conditionally unmapped by this
flag. This is all folded into the WasmNull to simplify RO space
iteration. In the future, we will make this unconditional, and will be
able to move the holes into the first page of the WasmNull.

Bug: 434179415
Change-Id: Ia9179e942cee45707c0a39c52d6d1312bcf4a9fd
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6965585
Auto-Submit: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102634}