[riscv][regexp] Use LoadAddress in PushBacktrack
This is much simpler and avoids emitting a weird forward label reference
in the middle of the instruction stream. The reference got patched when
the label was bound, but that meant that we would have a weird bound
label position embedded as a potentially illegal instruction mid stream.

Change-Id: I5ffd81cbbc7d50a48d5887806f35b21e14ab8c48
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6968858
Commit-Queue: Kasper Lund <kasperl@rivosinc.com>
Reviewed-by: Florian Loitsch <floitsch@rivosinc.com>
Cr-Commit-Position: refs/heads/main@{#102636}