[heap] Add more tracing output to incremental marking start
This CL adds more data to help understanding why incremental
marking is started. Apart from this, this CL is not supposed to
change any GC behavior.

This CL changes IncrementalMarkingLimitReached() to also return
the reason for the decision. When starting incremental marking
we emit this reason in a Perfetto event along with additional data
about current heap sizes.

This CL also changes ComputeNewAllocationLimits() to
UpdateAllocationLimits(). The method now applies the new limits
and emits a Perfetto event as well.

Change-Id: Iec529a0fc068e528c4db03fbc8cd788b33227090
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6965615
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102642}