[cleanup] Fix 3 ClangInliner findings:
The use of this symbol has been deprecated and marked for inlining. The
function being deprecated is absl::Mutex::Lock.

The use of this symbol has been deprecated and marked for inlining. The
function being deprecated is absl::Mutex::TryLock.

The use of this symbol has been deprecated and marked for inlining. The
function being deprecated is absl::Mutex::Unlock.

Bug: 394494905
Change-Id: I984b2861002ac7244422ad3a88c0325b502ded46
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6836428
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102656}