[maglev] Set Smi type if no hole loads
This is a bug introduced by https://crrev.com/c/6965618.

If we should treat hole as undefined when loading, we
cannot set the load type to Smi. Otherwise we may
create a deopt loop.

However, if we know the load cannot be a hole, ie,
we emit a BuildCheckNotHole, then if the elements
kind is a Smi, we must be loading a Smi.

Bug: 446106990
Change-Id: I7ec12626464b81d3f5e56aec4dfa20a7efb0da2c
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6967745
Auto-Submit: Victor Gomes <victorgomes@chromium.org>
Commit-Queue: Victor Gomes <victorgomes@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102659}