[compiler] Handle holes in heap ref Is-checks
When heap access is allowed, heap ref `IsFoo` previously called the
matching runtime `IsFoo` helper directly. This is incorrect for hole
values, since IsFoo might want to read the value's map.

This patch explicitly checks for holes in these predicates, and exits
the predicate early if a hole is found. This is unlike the runtime
`IsFoo` checks, since for those we want to explicitly check for a hole
earlier, for only cases where holes are allowed. In compilers, however,
a valid graph might have a hole flowing into a node where it is not
expected, since parts of the graph may be dead in terms of control flow,
yet still exist. This means that value checks in the compilers have to
be more robust against situations that are at runtime impossible.

Bug: 434179415
Change-Id: Iae9d8112d2d9bd4a75495506604e179785d92119
Fixed: 446141619
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6974219
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Auto-Submit: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102687}