[wasm] Reduce WasmNull size when not unmapping holes
The second half of the WasmNull payload is conditionally unmapped by
--unmap-holes, which means that it still needs to have valid hole maps
when the holes are not unmapped. This means that the holes need to be
iterated, at least by the RO space iteration during serialization.

We can make this work with the rest of the system a bit better by making
the WasmNull size dynamic, and dependent on the --unmap-holes flag, so
that with hole unmapping it runs over the holes, and without it the
holes are like normal.

Bug: 434179415
Change-Id: I61ca82913709cac8c72fab48ae46c413e8001e3b
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6973468
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Auto-Submit: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102703}