Sample PKU support when adding an isolate
We were recording a histogram sample for PKU support when creating the
first `NativeModule` in the isolate. As creating a `NativeModule` needs
to be isolate-independent in the future, we are removing this isolate
access and instead log the sample when adding the isolate to the engine.

Note that this changes the amount of samples; we now record a sample for
every isolate, even if it never uses WebAssembly.
If this turns out to skew the number on the histogram we will revisit
this later.

R=jkummerow@chromium.org

Bug: 446885278
Change-Id: I0fed070a9bcce643ba59baf877e1a5fdec173eed
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6973547
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102720}