[riscv] Fix trampoline pool next check position
Before this change, we would move the position of the trampoline pool's
next check forward when we tried to emit the pool while blocked. The
intent was to get the pool emitted when leaving the blocking scope, but
if we got unlucky and had a label bound inside the scope, we would end
up increasing the updated next check position even further as we removed
the adjustment we had for the unbound label. This, in effect, could make
us not emit the pool when leaving the blocking scope, thus potentially
making it impossible to reach from a label usage to the trampoline pool.

Change-Id: I66cfcfc3f391dc77cd4834592f7dd3993dc666cd
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6976253
Reviewed-by: Florian Loitsch <floitsch@rivosinc.com>
Commit-Queue: Kasper Lund <kasperl@rivosinc.com>
Cr-Commit-Position: refs/heads/main@{#102729}