[compiler] Mark holes as background serialized
Allocate a full HeapObjectData for the holes, and mark them as
background serialized, so that their map is cached. This allows the
compiler to read hole maps in optimisations and pattern matching, even
if the holes are unmapped.

Also some drive-by cleanup that replaces LookupOrInsert with InsertNew,
since we already did a lookup at the start of TryGetOrCreateData.

Bug: 434179415
Change-Id: Ifc5ebb7eaf1275a9282086f9b7eb9654e3d68991
Fixed: 446730212
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6979444
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102732}