cppgc: Mitigate Oilpan VA exhaustions
Currently, Oilpan tries to overreserve the cage to make sure that the
LSB of the masked out pointer part is 1. It does it by allocating twice
the actual useful reservation size and twice the needed alignment
(16GB). This causes the platform allocator to mmap 64GB - kPageSize,
which may often fail on systems with low reservation limits
(VA_BITS==39), especially with the presence of the V8 Sandbox and
the out-of-line PA metadata.

The CL aims to mitigate it by using the opportunistic strategy:
1) first, relax the requirement of the alignment to be twice the needed
   alignment. Instead, just pick the needed half. This will try to
   mmap 48GB - kPageSize.
2) if this fails, try to allocate the actual useful reservation size
   (16GB) for a few tries until the LSB of the masked out part is 1.

Bug: 444613347
Change-Id: Idea29fd019d75e9904648bcdcdde8754b3475ac1
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6973464
Commit-Queue: Anton Bikineev <bikineev@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102748}