[wasm] Add cache for ReduceWasmTypeCastRtt in loops
This CL adds a cache for ReduceWasmTypeCastRtt with exactness kExactMatchLastSupertype in loops. The cache is implemented using a
loop phi, itself introduced by using Variables from the
VariableReducer.

A new WasmTypeCastRttAnalyzer analyzer computes which WasmTypeCastRtt
are in which loop. Since this analyzer is a bit expensive (because it
uses a LoopFinder and because it analyzes the whole graph), this CL
also adds WasmLoopTypeCastRttFinder, which is a "pre-analyzer" that
runs as a Reducer during the phase before and pre-computes whether the graph contains WasmTypeCastRtt inside loops.

Change-Id: Icec7b41f23c3ff5bc429831be35df65a3e07fb65
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6965082
Commit-Queue: Nico Hartmann <nicohartmann@chromium.org>
Auto-Submit: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102768}