Introduce Relaxed_FetchOr for MarkBit::Set
Relaxed_SetBits were using Relaxed_CompareAndSwap to set the bits. It
contains a loop to ensure that the three operations (getting bits
/bitwise or/setting bits) are atomic. However, when all the bits to be
set are 1, we can use atomic fetch_or directly to avoid this loop.
This can be used in MarkBit::Set.

Change-Id: I197f79c99a6721157460544ddca108c1364d7aef
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6983342
Commit-Queue: Hao A Xu <hao.a.xu@intel.com>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102797}