[wasm] Add concept of delayed counter updates
When starting module compilation in the background we do not have access
to an isolate yet. For finishing compilation, we will later have that,
though.
So this CL refactors how the `NativeModule` and the
`WasmImportWrapperCache` (via the `WasmCodeAllocator`) do counter
updates. Instead of storing a shared pointer to one isolate's counters
(which could become stale, in particular for the import wrapper cache),
we do store outstanding counter updates in a separate data structure and
publish them the next time we have access to an isolate.

R=jkummerow@chromium.org

Bug: 446885278
Change-Id: I56644e5798cbd9171dd02f0bc84b94f922f145bc
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6983283
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102855}