[baseline] Make ForOfNextBaseline write the registers
ForOfNextBaseline can call user code, which can discard the baseline
code containing the ForOfNextBaseline code if it needs to resume in the
interpreter because someone set a breakpoint in that function.

If this happens, the StoreRegisterPair code is not executed, and the
done/value registers are not updated. This means that we'll still have
the old value for "done", which happens to be "true" and the iteration
will terminate.

Instead, we can make the ForOfNextBaseline builtin call the main
ForOfNext builtin, and then write the interpreter registers itself.

Additionally, mark the bytecode as clobbering the accumulator, so that
we don't need to save the accumulator across the call (which also
doesn't work with the deopts).

Bug: 408061015
Change-Id: I7cd786d633ec9b8c04b28cf628ce78b53297bd61
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6983293
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Rezvan Mahdavi Hezaveh <rezvan@chromium.org>
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102818}