Use dereference operator for `Optional` value access.
Replaces `new_map.value()` with `*new_map` after confirming
`new_map.has_value()`, which is the idiomatic way to access
the value of an `Optional`. The `value()` method combines two
actions: a test for validity followed by an access of the value,
which is redundant if the validity has already been checked.

Change-Id: I093e24e4252145d38a194a43572022570cc62f90
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6997690
Reviewed-by: Jakob Linke <jgruber@chromium.org>
Commit-Queue: Jeffrey Yu <yuje@google.com>
Reviewed-by: Thibaud Michaud <thibaudm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102852}