[riscv] Clear the vector unit at branch targets
The original code was clearing the vector unit at branch points and not
at the target labels. This could lead to cases where the unit wasn't
cleared correctly:

```
  VU.set(... m1)
  vmul(...)  # Use the vector with m1.
target:
  // Didn't set the vector to 'm1' as it is already at m1.
  vmul(...)  # Use the vector with m1.  <--- (1)

  VU.set(... m2)
  ...
  jge target  # Would now have the lmul at m2, instead of m1. <--- (2)
  VU.set(... m2)  // Would now set it again.  <--- (3)
  vmul(...)
```

The `vmul` at (1) wouldn't set the lmul, as it already was at m1.
However, that wouldn't be true if the branch at (2) was taken.

At the same time, clearing the vector unit at branch points would clear
it at (2), requiring it to be set again after the branch (3).

The new code clears the vector unit when a label is bound, or when a
call is made:

```
  VU.set(... m1)
  vmul(...)
target:  # Clears the unit. <--- (a)
  VU.set(... m1)  # Now sets it again. <--- (b)
  vmul(...)

  VU.set(... m2)
  ...
  jge target
  vmul(...)  # No need to set it again (c).
```

Since the unit is cleared at the bound label (a), it must be set again
afterwards (b). However, we can avoid setting it after a conditional
jump now (c).

Change-Id: I1c06a30a73a05d8886c6564a940d16c6ddef1c1a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6999410
Reviewed-by: Kasper Lund <kasperl@rivosinc.com>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Florian Loitsch <floitsch@rivosinc.com>
Cr-Commit-Position: refs/heads/main@{#102847}