[snapshot] Fix serialization of v8::External
v8::External is getting refactored to allow type-specific pointer tags,
i.e. that not all pointers stored within a v8::External have the same
tag.

For the serialization and deserialization of v8::External this means
that the pointer tags have to be written to the serialization format as
well.

Change-Id: I509bb8bf62ce8624435273c6690be34fa72b2f9f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6983087
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Jakob Linke <jgruber@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102845}