[regexp][x64] Fix SkipUntilOneOfMasked without SSE4.1
We had 1 pcmpeqd instruction were the destination register was not equal
to the first src register. Since pcmpeqd < SSE4.1 has only 2 registers,
this was invalid.

Drive-by: Fix overeager DCHECK about number of characters being 4. The
characters are the result of the input characters bitwise and the mask,
so 0x00 is a valid character that would cause this DCHECK to fail.

Fixed: 447647871
Bug: 437003349
Change-Id: Ie4f2d4b14a88e96ce4a53f083d518ce63c204911
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6999652
Reviewed-by: Jakob Linke <jgruber@chromium.org>
Auto-Submit: Patrick Thier <pthier@chromium.org>
Commit-Queue: Jakob Linke <jgruber@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102882}