[wasm] Remove explicit initialization of WasmImportWrapperCache
By just default-allocating the `WasmCodeAllocator`, it will reserve the
first code space when the first import wrapper is added.

This changes two things:
1) Instead of reserving 32kB initially, the `WasmCodeAllocator` will
   reserve a single page only (assuming the wrapper code fit within one
   page).
   So we might need to allocate 1-2 more code spaces if many wrappers
   are compiled in one process.
2) We do not have an isolate to trigger GC when allocation fails.
   This is unlikely to increase the number of OOMs, since
   a) we only had an isolate for the first 32k reservation, and not for
      code spaces added later, and
   b) import wrappers are rather small compared to actual Wasm code, so
      it's unlikely that we run out of memory when compiling import
      wrappers.

If we *do* see an increase in OOM, we could still pass the isolate to
`WasmImportWrapperCache::ModificationScope::AddWrapper` (all callers
have an isolate), and pass it on through
`WasmCodeAllocator::AllocateForWrapper` to
`WasmCodeAllocator::AllocateForCodeInRegion`, where we can trigger a GC
when a code space reservation fails.
We will leave that option for later though.

R=jkummerow@chromium.org

Bug: 446885278
Change-Id: Iea819f25171a1eebedaca093fc6f63340a69000f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6988709
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102869}