[wasm] Store canonical signature on WasmInternalFunction
We currently look up the signature from the module (using the stored
function index) when we need it. In the case of imports this can easily
go wrong though, since the stored function index is the import index,
but the stored `implicit_arg` can be a `WasmTrustedInstanceData` (of the
module that originally defined the function).

In order to avoid confusion there, we now store the signature directly
on the central `WasmInternalFunction`, and in turn remove the fields
from `WasmExportedFunctionData`, `WasmJSFunctionData`, and
`WasmCapiFunctionData` (which all link to an internal function anyway).

R=jkummerow@chromium.org

Bug: 432289371, 435582611
Change-Id: Ib7ddae0bd002c988d485b7d381df3b7a5c961a63
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6811187
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102899}