[wasm] Update grown shared memories without allocating
When a shared memory has grown, we need to allocate a new
JSSharedArrayBuffer for it. This patch moves that allocation
from the stack guard to the next invocation of the getter for
that buffer, which is the earliest time it becomes observable.

This is a prerequisite for avoiding GC on loop back edges.

Change-Id: I6a6e59a09f306692386925261f03453d5ccfd7f4
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7003085
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Commit-Queue: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102900}