[heap] Add additional tracing output when computing limits
This CL adds additional tracing output when computing new heap
allocation limits. For this purpose ComputeNewAllocationLimits()
starts setting the  new limits as well and because of that is
renamed to UpdateAllocationLimits().

To make this possible UpdateAllocationLimits() now accepts
boundaries as a parameter which callers can use to force the
new limits to be either at least as large (or at most) as the current
limits.

Bug: 408962793
Change-Id: Ide5140dc97c1ab490c539e705e20de2e140ced99
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7002926
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102895}