[heap] Reach safepoint without triggering GC for left-trimming
During left-trimming verification we need to reach a safepoint to
iterate roots safely. However, with a shared heap a global
safepoint may have already locked local_heaps_mutex_ to perform a
shared GC. In such a scenario a shared GC would run during left
trimming before the pointer in the owning JSArray got updated.

This CL fixes this by using ReachSafepointWithoutTriggeringGC()
to reach the safepoint. This method will always succeed to reach
a safepoint without a shared heap but it might fail when a shared GC
already holds the lock. Verification will only run when the safepoint
was reached. Even with a shared heap, we should reach the safepoint
in the vast majority of cases.

Bug: 446992085
Change-Id: I519fe9fe45f3a5b15cb3e87533907251e547516f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7002762
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102892}