[wasmfx] Fix handler validation
Handler validation was incomplete: the resume handlers should push the
captured continuation on the stack, and this should be reflected in the
target block signature.
Push the missing cont ref, and add all the type checks that apply to the
new continuation's signature.

R=clemensb@chromium.org
CC=fgm@chromium.org

Bug: 388533754
Change-Id: I79aa717beaacd3a119957b380a4e8e9d58f9c48f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7003021
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102896}