[wasm] Pass the ModuleOrigin if that's all we need
For creating export wrappers we need to know if the module is an asm.js
module, because the wrapper behaves slightly different in that case.

Instead of passing the whole WasmModule, just pass the origin. This
makes it more clear at call sites what this information might be used
for.

R=jkummerow@chromium.org

Bug: 435582611
Change-Id: I8d37a142f596fefc250d740d37c85e2df77d39e3
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7003273
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102909}