[deoptimizer] Fix transmitting DeoptimizeReasons
Before the fix, GetDeoptInfo returned an unrelated DeoptimizationReason
if there was no data associated with the queried_pc. This was because we
retrieved data for the nearest pc which was < queried_pc and had data.

This fix makes the pc matching exact, so we know whether there is or
isn't data associated with queried_pc.

Change-Id: I8cf5aa0ef62ec4931cca3b02b4adad34fbf60532
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7003387
Commit-Queue: Marja Hölttä <marja@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102911}