Reland [deoptimizer] Fix transmitting DeoptimizeReasons
Before the fix, GetDeoptInfo returned an unrelated DeoptimizationReason
if there was no data associated with the queried_pc. This was because we
retrieved data for the nearest pc which was < queried_pc and had data.

This fix makes the pc matching exact, so we know whether there is or
isn't data associated with queried_pc.

Last try: https://chromium-review.googlesource.com/c/v8/v8/+/7003387

Bug: 431933185
Change-Id: Id4af45bf66596bf3107485f7467858c13e91ed08
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7003400
Commit-Queue: Marja Hölttä <marja@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102939}