[test][wasm-js] Fix loading resources for proposal tests
Do not rely on the (outdated) `proposal_flags` to identify which test
belongs to a proposal. Instead look for a directory named "proposals"
near the end of the test directory.

This fixes one test failure which used the upstream
`wasm-module-builder.js` file instead of the copy from the proposal
repository.

Drive-by: Remove outdated proposal flag definitions.

R=mliedtke@chromium.org

Bug: 449581914
Change-Id: Ifcadc83da8d1eff0bbac7b3a69bb19406b5452f5
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7003200
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102956}