[heap] Remove OffHeapFullPointerSlot
The reason to have OffHeap- variants of compressed
pointer slots is to remove the operator*, which would
not work - it uses the location of the slot to work
out the base of the pointer cage.

However the operator* of a FullPointerSlot works just
fine, even off-heap, because it doesn't need any cage
base to decompress a pointer that is not compressed.

We were not consistently using OffHeapFullPointerSlot
anyway - VisitRoots was not using it even though
roots are off-heap.

Change-Id: I296385f7ee7022e8b9ad2a52770f458051f971ef
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7002994
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Erik Corry <erikcorry@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102954}