[ept] Delete tag kExternalObjectValueTag
The tag was a generic tag used for v8::External. After a recent
refactoring, the generic tag should not be used anymore. Instead,
type-specific tags should be used.

Additionally the number of tags reserved for v8::External gets reduced
from 100 to 40, as this is enough in currently for Chrome, where 34 tags
are used so far. The reduction is needed to make sure that the
ExternalPointerTags stay within 7 bit range.

Another change is that the tags that were used so far for tests are out
of range now. Therefore, the tags get shifted to a new value range.

Bug: 436799229
Change-Id: I0bbef6391bd4340d8d113ba01a565d9f1a24a0a4
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7003791
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102975}