[regexp] Make Bytecode operand value getters GC safe
The getter for BitTable previously returned a raw pointer to embedded
data within the Bytecode. This is not GC safe, as the pointer will
become stale whenever the Bytecode is moved during GC.

To prevent this, the existing value getters now require a
DisallowGarbageCollection scope.
In addition new variants of the getters are added that are GC-safe
(operate on and return handles).

Fixed: 449341185
Change-Id: I53f0e8d961bc916ca35228b74b2aad9253e7c69d
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7016609
Commit-Queue: Patrick Thier <pthier@chromium.org>
Reviewed-by: Jakob Linke <jgruber@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102983}