[arm64][builtins] Ensure remapping works on non-4k linux systems
This ensures that we align builtins to the largest possible size on
Linux (64k) and Android (16k), so that remapping them works regardless
of the runtime OS page size.

We should note that we cannot rely on the V8_OS_ANDROID macro as
targeting Android means we're cross-compiling. We have to rely on the
runtime `--target_os=android` flag instead.

Change-Id: I093fe8f4371b7e994f0b63a0ee7c6064e8f85747
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7013751
Reviewed-by: Jakob Linke <jgruber@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Pierre Langlois <pierre.langlois@arm.com>
Cr-Commit-Position: refs/heads/main@{#103157}