[heap] Remove uses of UNSAFE_SKIP_WRITE_BARRIER in C++ write barriers
This CL removes usages of UNSAFE_SKIP_WRITE_BARRIER for C++ write
barriers. We were using it only for avoiding write barriers on
InstructionStream but we can now avoid this with the
WriteBarrierModeScope. This CL uses that class to allow skipping
the write barrier on that InstructionStream without
UnsafeSkipWriteBarrier.

Bug: 437096305
Change-Id: Iaeaf934c466a0033bbc8dfd04eeba93a35037ad5
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7014096
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#102990}