[csa] Enforce valid Label for non-computed hashes in LoadNameHash
Remove the default argument for `if_hash_not_computed` in
CodeStubAssembler::LoadNameHash() and ensure that passed arguments are
not null.
This prevents accidental misuse of `LoadNameHash()` when either the hash
can be non-computed and that case needs to be dealt with by the caller,
or the hash is guaranteed to be computed (in that case
`LoadNameHashAssumeComputed` should be used).

Bug: 447954816
Change-Id: I1ef0fd2b53de6a7c4aabf5e807457e63ba0a86b5
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7020171
Commit-Queue: Toon Verwaest <verwaest@chromium.org>
Commit-Queue: Patrick Thier <pthier@chromium.org>
Reviewed-by: Toon Verwaest <verwaest@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103026}