[test] Fix shared string test on non-sandbox builds
test-shared-strings/ExternalizeSharedWithoutSharedStringTable added
in [1] fails on builds with pointer-compression enabled but sandbox
disabled, due to the string that is externalized being too short to be
in-place externalizable in this config.
Use a larger string for the test to fix this.

[1] https://crrev.com/c/7021469

Bug: 446981745
Change-Id: Icb394a63f3dd1264f11f2c38621c7a4673560055
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7022815
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Auto-Submit: Patrick Thier <pthier@chromium.org>
Commit-Queue: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103029}