[wasm] Prepare NativeModule in the background
The `PrepareAndStartCompile` phase does not need to be synchronous any
more. The main task is allocating the `NativeModule` and potentially
triggering compilation tasks. This can be done from the background,
removing work from the main thread's queue during startup where there is
a lot of stuff to do in the foreground already.

It also moves us closer to a new API which allows starting module
compilation from the background.

Bug: 446885278
Change-Id: Ia633d86ebd4d0f492c854af9029d1e09fc9e5bd1
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6973469
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#103067}